requirements.txt
djangorestframework==3.15.1
selenium==4.21.0
celery==5.4.0
requests==2.32.3
